Possibility to use images with src just like icons
To set an icon to your title or groupItem or item, you set its icon
parameter to a fontawesome icon string.
Now you can also use any image the same way with the image parameter.
This image paramater has to be the src string you want to put in the <img /> tag.
So it can be an absolute path (https://some-url.com/some-image.jpg) or a
relative path (/Images/some-image-on-my-website.jpg)